requirements.txt
selenium==4.18.0
webdriver-manager==4.0.1
pandas==2.1.4
sqlite3==3.45.1
apache-airflow==2.8.1
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.0
